% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{GLM}
\alias{GLM}
\title{Binding to the GLM.jl namespace}
\format{
An object of class \code{JuliaModuleImport} of length 149.
}
\usage{
GLM
}
\description{
Binding to the GLM.jl namespace
}
\keyword{datasets}
