% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlme-package.R
\docType{package}
\name{jlme-package}
\alias{jlme}
\alias{jlme-package}
\title{jlme: Regression Modelling with 'GLM.jl' and 'MixedModels.jl' in 'Julia'}
\description{
Bindings to 'Julia' packages 'GLM.jl' \doi{10.5281/zenodo.3376013} and 'MixedModels.jl' \doi{10.5281/zenodo.12575371}, powered by 'JuliaConnectoR'. Fits (generalized) linear (mixed-effects) regression models in 'Julia' using familiar model fitting syntax from R. Offers 'broom'-style data frame summary functionalities for 'Julia' regression models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://yjunechoe.github.io/jlme/}
  \item \url{https://github.com/yjunechoe/jlme/}
  \item Report bugs at \url{https://github.com/yjunechoe/jlme/issues/}
}

}
\author{
\strong{Maintainer}: June Choe \email{jchoe001@gmail.com} (\href{https://orcid.org/0000-0002-0701-921X}{ORCID})

}
\keyword{internal}
