% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{jlmer}
\alias{jlmer}
\title{Fit a mixed-effects regression model using MixedModels.jl}
\usage{
jlmer(formula, data, family = NULL, contrasts = jl_contrasts(data), ...)
}
\arguments{
\item{formula}{A formula written in Julia syntax}

\item{data}{A data frame}

\item{family}{A distribution family}

\item{contrasts}{A Julia dictionary of contrasts
Inferred from \code{data} by default.}

\item{...}{Additional options to \code{fit()}, called in Julia}
}
\value{
A julia model object of class \code{jlme}
}
\description{
Fit a mixed-effects regression model using MixedModels.jl
}
\examples{
stop_julia()
jlme_setup()
library(lme4)

lmer(Reaction ~ Days + (Days | Subject), sleepstudy)
jlmer(Reaction ~ Days + (Days | Subject), sleepstudy)
jlmer(Reaction ~ Days + (Days | Subject), sleepstudy, REML = TRUE)

glmer(r2 ~ Anger + Gender + (1 | id), VerbAgg, family = "binomial")
jlmer(r2 ~ Anger + Gender + (1 | id), VerbAgg, family = "binomial")

stop_julia()
}
