% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{jmer}
\alias{jmer}
\title{Run model in Julia, return to R}
\usage{
jmer(formula, data, REML = TRUE, family = NULL)
}
\arguments{
\item{formula}{model formula}

\item{data}{data set}

\item{REML}{use REML?}

\item{family}{(string)}
}
\description{
Run model in Julia, return to R
}
\examples{
setup()
data("sleepstudy", package="lme4")
data("InstEval", package="lme4")
library(lme4)
system.time(jm1 <- jmer(Reaction ~ Days + (Days|Subject), sleepstudy)) ## ~ 0.04 seconds
system.time(rm1 <- lmer(Reaction ~ Days + (Days|Subject), sleepstudy)) ## ~ 0.02 seconds
\dontrun{
system.time(rm2 <- jmer(y ~ 1 + service + (1|s) + (1|d) + (1|dept), InstEval, REML=FALSE))
system.time(jm2 <- jmer(y ~ 1 + service + (1|s) + (1|d) + (1|dept), InstEval, REML=FALSE))
}
}
