% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{MixedModels}
\alias{MixedModels}
\title{Binding to the MixedModels.jl namespace}
\format{
An object of class \code{JuliaModuleImport} of length 268.
}
\usage{
MixedModels
}
\description{
Binding to the MixedModels.jl namespace
}
\keyword{datasets}
