% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profilelikelihood.R
\name{profilelikelihood}
\alias{profilelikelihood}
\title{Profile the likelihood surface of Julia mixed effects models}
\usage{
profilelikelihood(x, ...)
}
\arguments{
\item{x}{A Julia MixedModel of class \code{jlme}}

\item{...}{Not implemented}
}
\value{
MixedModels.profile() output as object of class \code{jlmeprof}
}
\description{
Profile the likelihood surface of Julia mixed effects models
}
\examples{
\donttest{
jlme_setup(restart = TRUE)

jmod <- jlmer(Reaction ~ Days + (Days | Subject), lme4::sleepstudy)
tidy(jmod)

prof <- profilelikelihood(jmod)
prof

tidy(prof)

stop_julia()
}
}
