% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{jlm}
\alias{jlm}
\title{Fit a fixed-effects regression model using GLM.jl}
\usage{
jlm(formula, data, family = NULL, contrasts = jl_contrasts(data), ...)
}
\arguments{
\item{formula}{A formula written in Julia syntax}

\item{data}{A data frame}

\item{family}{A distribution family}

\item{contrasts}{A Julia dictionary of contrasts
Inferred from \code{data} by default.}

\item{...}{Additional options to \code{fit()}, called in Julia}
}
\value{
A julia model object of class \code{jlme}
}
\description{
Fit a fixed-effects regression model using GLM.jl
}
\examples{
stop_julia()
jlme_setup()

lm(mpg ~ hp, mtcars)
jlm(mpg ~ hp, mtcars)

x <- mtcars
x$cyl_sum <- factor(x$cyl)
contrasts(x$cyl_sum) <- contr.sum(3)
lm(mpg ~ cyl_sum, x)
jlm(mpg ~ cyl_sum, x)

x$cyl_helm <- factor(x$cyl)
contrasts(x$cyl_helm) <- contr.helmert(3)
colnames(contrasts(x$cyl_helm)) <- c("4vs6", "4&6vs8")
lm(mpg ~ cyl_helm, x)
jlm(mpg ~ cyl_helm, x)

stop_julia()
}
