% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlme-package.R
\docType{package}
\name{jlme-package}
\alias{jlme}
\alias{jlme-package}
\title{jlme: Regression modelling with 'GLM.jl' and 'MixedModels.jl'}
\description{
Bindings to 'Julia' packages 'GLM.jl' and 'MixedModels.jl'.
}
\author{
\strong{Maintainer}: June Choe \email{jchoe001@gmail.com} (\href{https://orcid.org/0000-0002-0701-921X}{ORCID})

}
\keyword{internal}
