% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{jlme_tidiers}
\alias{jlme_tidiers}
\alias{tidy.jlme}
\alias{glance.jlme}
\title{Tidier methods for Julia regression models}
\usage{
\method{tidy}{jlme}(x, effects = c("var_model", "ran_pars", "fixed"), ...)

\method{glance}{jlme}(x, ...)
}
\arguments{
\item{x}{An object of class \code{jlmer_mod}}

\item{effects}{One of "var_model", "ran_pars", or "fixed"}

\item{...}{Unused}
}
\value{
A data frame
}
\description{
Tidier methods for Julia regression models
}
